+++ beaker/conftest.py
@@ -9,17 +9,21 @@ def doctest_fixtures(
     docker_image_name,
     beaker_image_name,
     beaker_cluster_name,
+    beaker_on_prem_cluster_name,
     experiment_name,
     dataset_name,
     download_path,
     beaker_org_name,
+    beaker_node_id,
 ):
     doctest_namespace["beaker"] = client
     doctest_namespace["workspace_name"] = workspace_name
     doctest_namespace["docker_image_name"] = docker_image_name
     doctest_namespace["beaker_image_name"] = beaker_image_name
     doctest_namespace["beaker_cluster_name"] = beaker_cluster_name
+    doctest_namespace["beaker_on_prem_cluster_name"] = beaker_on_prem_cluster_name
     doctest_namespace["experiment_name"] = experiment_name
     doctest_namespace["dataset_name"] = dataset_name
     doctest_namespace["download_path"] = download_path
     doctest_namespace["beaker_org_name"] = beaker_org_name
+    doctest_namespace["beaker_node_id"] = beaker_node_id